requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
jinja2
python-dotenv
